+++ Sources/SwiftyChart/Barchart/BarChartModel.swift
@@ -9,13 +9,13 @@ import Foundation
 import SwiftUI
 
 public struct BarChartModel: ChartModel {
-    public init(value: CGFloat, color: Color, title: String) {
+    public init(value: Float, color: Color, title: String) {
         self.value = value
         self.title = title
         self.color = color
     }
     
-    public var value: CGFloat
+    public var value: Float
     public var title: String
     public var color: Color
 }
